+++ py/bitbox02/bitbox02/bitbox02/bitbox02.py
@@ -212,6 +212,17 @@ class BitBox02(BitBoxCommonAPI):
             raise
         return True
 
+    def change_password(self) -> None:
+        """
+        Changes the device password. The user must unlock with the old password
+        and enter and confirm the new password.
+        Raises a Bitbox02Exception on failure.
+        """
+        # pylint: disable=no-member
+        request = hww.Request()
+        request.change_password.CopyFrom(bitbox02_system.ChangePasswordRequest())
+        self._msg_query(request, expected_response="success")
+
     def create_backup(self) -> bool:
         """
         Returns True if the backup was created successfully.