+++ src/rust/bitbox02-rust/src/shiftcrypto.bitbox02.rs
@@ -220,6 +220,9 @@ pub struct SetPasswordRequest {
 }
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, Copy, PartialEq, ::prost::Message)]
+pub struct ChangePasswordRequest {}
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, Copy, PartialEq, ::prost::Message)]
 pub struct BluetoothToggleEnabledRequest {}
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, Copy, PartialEq, ::prost::Message)]
@@ -1975,7 +1978,7 @@ pub struct Success {}
 pub struct Request {
     #[prost(
         oneof = "request::Request",
-        tags = "2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 24, 25, 26, 27, 28, 29"
+        tags = "2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 24, 25, 26, 27, 28, 29, 30"
     )]
     pub request: ::core::option::Option<request::Request>,
 }
@@ -2040,6 +2043,8 @@ pub mod request {
         Bip85(super::Bip85Request),
         #[prost(message, tag = "29")]
         Bluetooth(super::BluetoothRequest),
+        #[prost(message, tag = "30")]
+        ChangePassword(super::ChangePasswordRequest),
     }
 }
 #[allow(clippy::derive_partial_eq_without_eq)]